Issue-17 Refactor code
