+++ projects/server/src/graphql/dataSources/Category.ts
@@ -1,6 +1,6 @@
 import Pagination from './Pagination'
 import type { PaginationParams } from './Pagination'
-import type { PrismaClient, Category } from '../../../generated/client'
+import type { PrismaClient, Category, Prisma } from '../../../generated/client'
 import type { CreateCategoryInput, UpdateCategoryInput, DeleteCategoryInput } from '../../../generated/resolvers'
 import type { Maybe } from '../../../../shared/models'
 
@@ -12,7 +12,7 @@ export default class CategoryDataSource extends Pagination<Category> {
     this.prisma = prisma
   }
 
-  getCategory = async (id: string | number, options?: any) => {
+  getCategory = async (id: string | number, options?: Prisma.CategoryFindUniqueArgs) => {
     const prismaId = typeof id === 'string' ? parseInt(id, 10) : id
     const category = await this.prisma.category.findUnique({
       where: {
@@ -39,7 +39,6 @@ export default class CategoryDataSource extends Pagination<Category> {
       this.prisma.category.count(),
       this.prisma.category.findMany(this.findManyOptions),
     ])
-    console.log(categoryList)
     this.nodes = categoryList
     const response = this.generatePaginationResponse()
     return {
