+++ projects/server/src/index.ts
@@ -4,8 +4,6 @@ import mercurius from 'mercurius'
 import schema, { dataSources } from './graphql'
 import Prisma from './utils/Prisma'
 
-import type { Loaders } from './utils/types'
-
 const server = fastify({
   // logger: true,
 })
@@ -18,10 +16,6 @@ const contextBuilder = async () => {
 }
 type PromiseType<T> = T extends PromiseLike<infer U> ? U : T
 export type Context = PromiseType<ReturnType<typeof contextBuilder>>
-declare module 'mercurius' {
-  interface MercuriusContext extends Context {}
-  interface MercuriusLoaders extends Loaders {}
-}
 
 server.register(mercurius, {
   schema,
@@ -29,11 +23,11 @@ server.register(mercurius, {
   context: contextBuilder,
 })
 
-server.listen(4000, (err, address) => {
+server.listen(4000, (err) => {
   if (err) {
     server.log.error(err)
     process.exit(1)
   }
-  console.info('server listening on 4000')
+  console.info('\x1b[35mserver listening on 4000')
   server.log.info('server listening on 4000')
 })
